Added the inGame View, connected it to events, (but not to database), added some number checks

+++ Utilities/MemberNumberCheck.cs
@@ -14,7 +14,7 @@ namespace CAA_Event_Management.Utilities
 
             try
             {
-                int numberCheck1 = Convert.ToInt32(memberNumber);
+                double numberCheck1 = Convert.ToDouble(memberNumber);
             }
             catch
             {
